Turn legalhold requests into polysemy action
